requirements.txt
fastapi
uvicorn
pymc
scipy
sqlalchemy
python-dotenv
requests
websockets
fastapi[all]
psycopg2-binary
pymc
numpy
pandas
